Cardset handing and language filtering
CardSets filtered by language and general tidyup of code. App now handles first time users who haven't previously purchased a cardset to get one for free, but only cardsets that are annoted as FreePurchase.

Language filtering fixed in Select Cardset so that a language has to be selected, if turning off the only language chosen then that language is turned back on.